+++ libarchive/archive_write_set_format_mtree.c
@@ -389,13 +389,13 @@ dir_len(struct mtree_entry *me)
 		else if (*p == '/')
 			al = p - path;
 	}
-	if (l == -1)
+	if (l == (size_t)-1)
 		goto alen;
 	size = p - path;
 	rp = p = path;
 	while (*p != '\0') {
 		l = mbtowc(&wc, p, size);
-		if (l == -1)
+		if (l == (size_t)-1)
 			goto alen;
 		if (l == 1 && (wc == L'/' || wc == L'\\'))
 			rp = p;
@@ -404,7 +404,7 @@ dir_len(struct mtree_entry *me)
 	}
 	return (rp - path + 1);
 alen:
-	if (al == -1)
+	if (al == (size_t)-1)
 		return (0);
 	return (al + 1);
 }
